Adjusts column math
Hey @jmaddi -- really good question.

I've been playing around with this for the past half hour and it
*seems* impossible to exactly match the original. However, I've
adjusted the math to match the original as close as possible. Thanks
for the heads up!!!

I guess the issue comes down to which is most, the column width, or the
margin between the columns.